+++ konst/src/lib.rs
@@ -274,6 +274,14 @@
 //! - `"deref_raw_in_fn"` (disabled by default):
 //! Requires Rust nightly. Enables `const fn`s that need to dereference raw pointers.
 //!
+//! - `"mut_refs"`(disabled by default):
+//! Enables const functions that take mutable references.
+//! Use this whenever mutable references in const contexts are stabilized.
+//! Also enables the `"deref_raw_in_fn"` feature.
+//!
+//! - `"nightly_mut_refs"`(disabled by default):
+//! Enables the `"mut_refs"` feature. Requires Rust nightly.
+//!
 //! # No-std support
 //!
 //! `konst` is `#![no_std]`, it can be used anywhere Rust can be used.
@@ -304,6 +312,7 @@
     feature = "constant_time_slice",
     feature(const_slice_from_raw_parts, const_fn_union, const_ptr_offset)
 )]
+#![cfg_attr(feature = "nightly_mut_refs", feature(const_mut_refs))]
 #![cfg_attr(feature = "docsrs", feature(doc_cfg))]
 #![no_std]