+++ konst_macro_rules/src/lib.rs
@@ -1,6 +1,7 @@
 //! Implementation detail of the `konst` crate.
 #![no_std]
 #![cfg_attr(feature = "deref_raw_in_fn", feature(const_fn_union))]
+#![cfg_attr(feature = "nightly_mut_refs", feature(const_mut_refs))]
 
 #[macro_use]
 mod option_macros_;